+++ packages/components/alert/stories/alert.stories.tsx
@@ -189,7 +189,7 @@ const CustomStylesTemplate = (args) => {
           color={color}
           title="The documents you requested are ready to be viewed"
         >
-          <div className="flex items-center gap-1 mt-3">
+          <div className="flex items-center gap-1">
             <Button
               className="bg-background text-default-700 font-medium border-1 shadow-small"
               size="sm"
@@ -223,6 +223,14 @@ export const WithDescription = {
   args: {
     ...defaultProps,
     description: "This is an alert with a description",
+  },
+};
+
+export const WithoutIconWrapper = {
+  render: Template,
+  args: {
+    ...defaultProps,
+    hideIconWrapper: true,
   },
 };