+++ packages/components/drawer/src/index.ts
@@ -1,9 +1,24 @@
-import {ModalHeader, ModalBody, ModalFooter, ModalContent} from "@nextui-org/modal";
+import {
+  ModalHeader,
+  ModalBody,
+  ModalFooter,
+  ModalContent,
+  type ModalContentProps,
+  type ModalHeaderProps,
+  type ModalBodyProps,
+  type ModalFooterProps,
+} from "@nextui-org/modal";
 
 import Drawer from "./drawer";
 
 // export types
 export type {DrawerProps} from "./drawer";
+export type {
+  ModalContentProps as DrawerContentProps,
+  ModalHeaderProps as DrawerHeaderProps,
+  ModalBodyProps as DrawerBodyProps,
+  ModalFooterProps as DrawerFooterProps,
+};
 
 // export hooks
 export {useDrawer} from "./use-drawer";